requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-datatables-view==1.20.0
django-js-asset==2.2.0
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
